requirements.txt
speechrecognition
pyttsx3
pyaudio
python-dotenv
openai
